[FE] #1-3 feat: answerUpdate(204)

+++ frontend-web/src/router/Main/index.jsx
@@ -2,12 +2,13 @@ import { useEffect, useState } from "react";
 import { Wrapper } from "./styles";
 
 import { useRecoilState } from "recoil";
-import { ChildrenID, MemberID, CurrentSlide, ChildrenList } from "../../atom";
+import { ChildrenID, MemberID, CurrentSlide, ChildrenList, NowAnswer, Chats } from "../../atom";
 
 import axios from "axios";
 import drf from "../../api/drf";
 import { Link } from 'react-router-dom';
 import MainCarousel from "../../components/Main/MainCarousel";
+import AnswerModal from "../../components/Childern/AnswerModal";
 
 
 function Main() {
@@ -15,9 +16,11 @@ function Main() {
   const [childrenID, setChildrenID] = useRecoilState(ChildrenID);
   const [currentSlide, setCurrentSlide] = useRecoilState(CurrentSlide);
   const [childrens, setChildrens] = useRecoilState(ChildrenList);
+  const [nowAnswer, setNowAnswer] = useRecoilState(NowAnswer);
 
-  const [children, setChildren] = useState([]);
-  // const [childrenNumber, setChildrenNumber] = useState(0);
+  const [isUpdating, setIsUpdating] = useState(false);
+
+  // const [children, setChildren] = useState([]);
 
   useEffect(() => {
     axios({
@@ -27,34 +30,29 @@ function Main() {
     }).then((res) => {
       setmemberID(res.data.memberID)
       setChildrenID(res.data.childrens[currentSlide]?.childrenID)
-      setChildren(res.data.childrens)
+      // setChildren(res.data.childrens)
       setChildrens([...res.data.childrens, {childrenId: 0}])
-      // setChildrenNumber(res.data.childrens.length)
     })
   }, [])
 
+  setChildrenID(childrens[currentSlide]?.childrenID)
   const answers = childrens ? childrens[currentSlide]?.answers : null;
-  // console.log(childrens[currentSlide])
 
   function deleteChildren(childrenID) {
-    axios({
-      url: drf.children.children(childrenID),
-      method: "delete",
-      headers: {Authorization: 'Bearer ' + localStorage.getItem("token"),},
-    }).then((res) => {
-      console.log(res)
-      window.location.reload()
-    })
-    .catch((err) => {console.log(err)})
+    if (window.confirm('정말 삭제하시겠습니까?')) {
+      axios({
+        url: drf.children.children(childrenID),
+        method: "delete",
+        headers: {Authorization: 'Bearer ' + localStorage.getItem("token"),},
+      }).then((res) => {
+        console.log(res)
+        window.location.reload()
+      })
+      .catch((err) => {console.log(err)})
+    }
   }
 
-  function updateAnswer(answerID, questionID) {
-    axios({
-      url: drf.answer.updateAnswer(answerID),
-      method: "put",
-      headers: {Authorization: 'Bearer ' + localStorage.getItem("token"),},
-    })
-  }
+  
 
   function deleteAnswer(answerID, questionID) {
     axios({
@@ -66,6 +64,29 @@ function Main() {
       window.location.reload()  // 새로고침 필요
     })
     .catch((err) => {console.log(err)})
+  }
+
+  function onSubmit(event) {
+    event.preventDefault();
+    const chat = event.target[0].value;
+    console.log(chat)
+  }
+
+  const [close, setClose] = useState(false) // 모달창 닫는 변수
+  const [num, setNum] = useState();
+
+  function updateActivate(answer, idx) {
+    console.log(answer)
+    setNum(idx)
+    // setIsUpdating(true)
+    setNowAnswer(answer)
+    setClose(!close)
+  }
+
+  console.log(num)
+
+  function updateAnswer(event) {
+    console.log(event.target)
   }
 
   return (
@@ -92,18 +113,17 @@ function Main() {
             </div>
           </div>
           <div className="body_grid">
-            { childrenID 
-              ? <button>추가하기</button>
-              : null
-            }
-            <p>\아이에게 해주고싶은말</p>
+          { childrenID ? <Link to={`/CreateAnswer/${childrenID}`}><button>추가하기</button></Link> : null }
+            <p>\아이에게 해주고싶은 말</p>
             <div className="body_conversation">
-              { answers ? answers.map((answer) => {
+              { answers ? answers.map((answer, idx) => {
                 return (
                   <div key={answer.answerID}>
-                    <span key={answer.answerID}>{answer.content}</span>
-                    <button onClick={() => updateAnswer(answer.answerID, answer.questionID)}>수정</button>
+                    <span>{answer.content}</span>
+                    <span>{answer.createdAt}</span>
+                    <button onClick={() => updateActivate(answer, idx)}>수정</button>
                     <button onClick={() => deleteAnswer(answer.answerID, answer.questionID)}>삭제</button>
+                    { close && num===idx && (<AnswerModal answer={answer} closeModal={() => setClose(!close)} />)}
                   </div>
                 )
               }) : null}
